fixed folder structure and removed unnecessary imports
