+++ resq/frontend/src/pages/UserRoles.js
@@ -8,7 +8,7 @@ import Box from '@mui/material/Box';
 import Grid from '@mui/material/Grid';
 import Typography from '@mui/material/Typography';
 import {createTheme, ThemeProvider} from '@mui/material/styles';
-import disasterImage from '../disaster.png';
+import disasterImage from './disaster.png';
 import Container from '@mui/material/Container';
 import Radio from '@mui/material/Radio';
 import RadioGroup from '@mui/material/RadioGroup';
@@ -16,14 +16,16 @@ import RadioGroup from '@mui/material/RadioGroup';
 
 function Copyright(props) {
     return (
-        <Typography variant="body2" color="text.secondary" align="center" {...props}>
-            {'Copyright © '}
-            <Link color="inherit" href="https://github.com/bounswe/bounswe2023group1">
-                ResQ
-            </Link>{' '}
-            {new Date().getFullYear()}
-            {'.'}
-        </Typography>
+        <div style={{position: 'fixed', bottom: 0, width: '100%'}}>
+            <Typography variant="body2" color="text.secondary" align="center" {...props}>
+                {'Copyright © '}
+                <Link color="inherit" href="https://github.com/bounswe/bounswe2023group1">
+                    <span style={{fontWeight: 'bold'}}>ResQ</span>
+                </Link>{' '}
+                {new Date().getFullYear()}
+                {'.'}
+            </Typography>
+        </div>
     );
 }
 
@@ -42,6 +44,7 @@ export default function SignIn() {
     const handleRoleChange = (event) => {
         setSelectedRole(event.target.value);
     };
+
     const handleSubmit = (event) => {
         event.preventDefault();
         const data = new FormData(event.currentTarget);
@@ -56,21 +59,20 @@ export default function SignIn() {
                 <CssBaseline/>
                 <Box
                     sx={{
-                        marginTop: 8,
+                        marginTop: 1,
                         display: 'flex',
                         flexDirection: 'column',
                         alignItems: 'center',
                     }}
                 >
-                    <Avatar sx={{width: 100, height: 100}}>
+                    <Avatar sx={{width: 80, height: 80, marginBottom: '10px'}}>
                         <img
                             src={disasterImage}
                             alt="Disaster"
                             style={{width: '100%', height: '100%', objectFit: 'cover'}}
                         />
                     </Avatar>
-
-                    <Typography component="h5" variant="h5" sx={{color: 'red', fontWeight: 'bold'}}>
+                    <Typography component="h5" variant="h5" sx={{color: 'red', fontWeight: 'bold', margin: '0'}}>
                         ResQ
                     </Typography>
                     <Box component="form" onSubmit={handleSubmit} noValidate sx={{mt: 1}}>
@@ -121,10 +123,9 @@ export default function SignIn() {
                         </Button>
                     </Box>
                 </Box>
-                <Copyright sx={{mt: 8, mb: 4}}/>
             </Container>
+            <Copyright sx={{mt: 5}}/>
         </ThemeProvider>
     );
 }
-
